fix status parsing for renames
in a rename, git status -z will print the two filenames (current and
previous) on two separate "lines" (delimited by nulls). the first "line"
has a status code, the second one does not. depending on that file's
name, it might be mistakenly parsed as a status code itself. we should
skip over that line to avoid any such errors.