move rate parameters to startx line
see https://bugzilla.redhat.com/show_bug.cgi?id=601853 - basically, when
udev detects a new keyboard, xorg will add that device to the input list
and this resets the rate options. adding a udev trigger will race with
xorg itself, so the best way is to set session-wide defaults for these
values. that way, when xorg resets them, it goes back to the defaults
that we defined earlier.