set st title
this is particularly important for some programs (ranger, cmus) that
want to set the title themselves. since they want to set it, we need to
put it back when we regain control.

also, just in general, having some information there is nice. the pwd is
usually enough for me to distinguish terminals in the rofi list.