add ranger rc
this program is tough to configure. especially the key mappings are all
over the place. i can see the vim inspiration but then there are lots of
things that don't make sense. some day i'll take a chainsaw to this and
tweak all the mappings to my liking. for now though this is sufficient