lock on key release
if you lock on key press, then the releasing of the key will often cause
the x server to wake from dpms off, which makes the -d switch to i3lock
useless. instead we should lock when the key is let go, so that the
extra key events are less likely to collide with the dpms off.

note that releasing the super key can also cause the monitor to turn
back on.