always open mpv window if x is available
i don't want to block all of ranger while listening to music, but if the
music has no album art, then it won't open a window, and since it was
forked, i can't access mpv's terminal ui either. let's always attach a
window for mpv so that we can access it in these situations.