add mpv replaygain support
replaygain is implemented as an audio filter that adds software volume
control. we have to set the defaults for the filter and then add the
filter to the default filter set. note that we do not use the detach
option to the filter, because this breaks volume control for files that
don't have replaygain metadata.