enable mpd socket activation
for some reason, mpd has systemd unit files for socket activation
system-wide, but not user-only. i am piggybacking on the existing
user/mpd.service included in mpd, and only adding the socket file, which
is only slightly modified from the upstream system/mpd.socket.

as of writing, there are only two differences between the user and
system versions of mpd.service:
- the system version has ProtectKernelModules (a comment in the user
  version clearly explains why this is not included there)
- the system version has Also=mpd.socket (the user version normally does
  not have the socket file so it doesn't include that option)
i have decided that the trouble of adding the Also= line is not worth
the hassle of diverging from the upstream user/mpd.service file, so i'm
just going to keep using it and only have a custom socket file.