fix maim bindings with -p -2
so maim doesn't like it when the borders of the window are offscreen.
that's bad because my bspwmrc does that by design. we can use slop to
select the window, identify its geometry, and then apply negative
padding to trim the borders off with the -p flag. however, maim with the
-i flag does not invoke slop, so the -p flag gets ignored. we'll have to
just give up on that binding.