fix duplicate rofi binding
went from rofi 1.3.1 to 1.4.2 today. this jump was drastic and brought
some significant config changes. it seems there is now another default
command bound to C-u so this command (which was bound to C-u by default
when i wrote this config originally) is no longer usable here.

i actually need to rewrite the entire config file for their new css-like
format introduced in 1.4.0, but i'm not up for that right now. soon.