rewrite and update ranger config
i had long planned to completely overhaul this configuration, but man
what a mess. ranger implements its own pager, line editor, and task
manager, all of which are sort of like the unixy tools you're used to
but not quite, and all with limited customization because they're
homemade. not great. maybe i should move to dired after all.