note export in bashrc
osx blurs the distinction between bash_profile and bashrc because it
doesn't load the profile when you log into the system. as a result,
terminal emulators in osx normally spawn a login shell, which is why
dotfiles on osx tend to be all over the place with where variables get
exported.

i have strong opinions on bash dotfile hygiene, and prefer to handle
exports in the profile where they can be inherited by child shells. so
we should move this somewhere else.