port bash tty ancestor check to fuser
arch released lsof 4.90 today. the last release of lsof was 4.89 from
july 2015. it seems that lsof learned how to detect linux pseudoterminal
endpoints in this release. if you ask "who has /dev/pts/0" open, lsof
will detect that /dev/pts/0 is a pty whose other endpoint is /dev/ptmx,
and it will also list processes that have /dev/ptmx open.

unfortunately, /dev/ptmx is the root pty node that all terminal
emulators use to launch a new pty device. so if you lsof /dev/pts/0,
you'll see your bash instance as well as the terminal emulator that it's
running in. this is explicitly the opposite of what i wanted lsof to do.
i only wanted to list the actual terminal applications.

you can approximately disable this behavior with -E, but -E does not
play well with another lsof flag i use, -c. frankly, -c has always been
kind of broken (try "lsof -Vc ^lsof", you'll see lsof complains about
being unable to find the command "^lsof"), so i'm not surprised that
stuff is borked.

the solution to this whole mess is to move to fuser, which is
posix-compliant, and not very smart. it has much more predictable output
and behavior, which is exactly what this problem needs. terminal devices
are hard!