always tile emacs
i used to be on the emacs-mac port, which (among other things)
automatically tiled nicely with chunkwm. however, the behavior of the
frame depends on the menu-bar-lines variable, so if i switch to a
desktop, osx will not refocus emacs automatically unless i have
menu-bar-mode on. in addition it seemed to have some sporadic init.el
loading issues.

so i switched back to brew's default emacs with cocoa enabled. so far it
seems to be better-behaved, but it doesn't tile by default, so i need a
rule to force it.

https://github.com/railwaycat/homebrew-emacsmacport

https://github.com/Homebrew/homebrew-core/blob/master/Formula/emacs.rb