disable tty check on osx
as of 10.13.6, running fuser for my tty (which actually invokes lsof)
takes 15+ seconds. i can reduce it to around 10s if i invoke lsof
directly with certain flags, but it's still unacceptable.

recall that, a long time ago, i had this functionality set to run on
every PROMPT_COMMAND, but it was too slow to tolerate on osx, so i moved
it to running at shell startup. well apparently apple has made it so
slow that now i can't tolerate it at all. might as well just turn it off
for this system.