remove firefox bspwm rule
this rule has done nothing but surprise and confuse me in the however
many years i've had it. whenever i move a firefox window to another
desktop, i end up with random dialog boxes and prompts from that window
going back to the original desktop, because of this rule. it needs to go