use environment.d
pam_env for users is deprecated, because pam itself is running as root
and doesn't want to be in the game of parsing user-owned files.
unfortunately, it was the only way to set envars uniformly.

systemd's environment.d directory sets all the envars for your systemd
user instance, but your user session doesn't run under the user systemd
so it doesn't pick up those vars. you have to load them yourself in your
shell files. this also means runuser will not pick up your envars unless
you run it with --login, which is a real shame. oh well, it looks like
we're not likely to see defragmentation in this problem space.